update with email settings

+++ application/models/student_model.php
@@ -62,18 +62,22 @@ class Student_Model extends CI_Model{
 
         try{
 
-            foreach ($language_set as $key => $value) {
-                    $this->db->where('id', $value['id']);
-                    $this->db->delete('user_language_set'); 
+            if (!empty($language_set)) {
+                foreach ($language_set as $key => $value) {
+                        $this->db->where('id', $value['id']);
+                        $this->db->delete('user_language_set'); 
+                }
             }
 
-            foreach ($profile['language'] as $key => $value) {
-                $userLanguage = array(  'user_id' => $this->session->userdata('id'),
-                                        'language' => $value['name'],
-                                        'written' => $value['written'],
-                                        'spoken' => $value['spoken'] );
+            if (!empty($profile['language'])) {
+                foreach ($profile['language'] as $key => $value) {
+                    $userLanguage = array(  'user_id' => $this->session->userdata('id'),
+                                            'language' => $value['name'],
+                                            'written' => $value['written'],
+                                            'spoken' => $value['spoken'] );
 
-                $this->db->insert('user_language_set', $userLanguage);                
+                    $this->db->insert('user_language_set', $userLanguage);                
+                }
             }
 
             if ($checkUserProfilePhotoExist) {
